backend/requirements.txt
xrpl-py>=4.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
websockets>=12.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
